update: logging on error

+++ app/Http/API/Account/UserController.php
@@ -6,6 +6,7 @@ use App\Helper\Constants;
 use App\Http\Controllers\BaseAPIController as Controller;
 use App\Http\Repositories\UserRepository;
 use App\Http\Response\BodyResponse;
+use App\Http\Response\ResponseCode;
 use Exception;
 use Illuminate\Http\Request;
 
@@ -24,16 +25,17 @@ class UserController extends Controller
         $this->repository = $repo;
     }
 
-    public function checkPermission($rule)
+    public function checkPermission($rule): bool|BodyResponse
     {
         try {
             if (!$this->repository->currentAccount()
                 ->hasPermissionTo($rule))
                 throw new Exception('Permission denied');
+            return true;
         } catch (\Throwable $th) {
             $body = new BodyResponse();
             $body->setPermissionDenied();
-            return $this->sendResponse($body);
+            return $body;
         }
     }
 
@@ -50,12 +52,22 @@ class UserController extends Controller
      */
     public function index(Request $request)
     {
-        $this->checkPermission($this->permissionRule()->index);
+        $checkPermission =$this->checkPermission($this->permissionRule()->index);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $order = $request->order ?? 'desc';
         $columns = $request->columns ?? ['*'];
         $count = $request->perPage ?? 0;
         $result = $this->repository->indexWithProfile($order, $request->all(), $columns, $count);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -66,12 +78,22 @@ class UserController extends Controller
      */
     public function indexTrashed(Request $request)
     {
-        $this->checkPermission($this->permissionRule()->index_trashed);
+        $checkPermission =$this->checkPermission($this->permissionRule()->index_trashed);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $order = $request->order ?? 'desc';
         $columns = $request->columns ?? ['*'];
         $count = $request->perPage ?? 0;
         $result = $this->repository->indexWithProfileTrashed($order, $request->all(), $columns, $count);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -83,8 +105,19 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $this->checkPermission($this->permissionRule()->store);
+        $checkPermission =$this->checkPermission($this->permissionRule()->store);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
+
         $result = $this->repository->create($request->all());
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -96,9 +129,19 @@ class UserController extends Controller
      */
     public function show(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->show);
+        $checkPermission =$this->checkPermission($this->permissionRule()->show);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->get('id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -110,9 +153,19 @@ class UserController extends Controller
      */
     public function showTrashed(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->show_trashed);
+        $checkPermission =$this->checkPermission($this->permissionRule()->show_trashed);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->getTrashed('id', $id);
+        
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -125,9 +178,19 @@ class UserController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->update);
+        $checkPermission =$this->checkPermission($this->permissionRule()->update);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->updateBy($request->all(), 'id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -140,9 +203,19 @@ class UserController extends Controller
      */
     public function restore(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->restore);
+        $checkPermission =$this->checkPermission($this->permissionRule()->restore);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->restoreBy('id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -154,9 +227,19 @@ class UserController extends Controller
      */
     public function destroy(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->destroy);
+        $checkPermission =$this->checkPermission($this->permissionRule()->destroy);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->deleteBy('id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -168,9 +251,19 @@ class UserController extends Controller
      */
     public function permanentDestroy(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->permanent_destroy);
+        $checkPermission =$this->checkPermission($this->permissionRule()->permanent_destroy);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->permanentDeleteBy('id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 
@@ -182,9 +275,19 @@ class UserController extends Controller
      */
     public function reset(Request $request, string $id)
     {
-        $this->checkPermission($this->permissionRule()->reset);
+        $checkPermission =$this->checkPermission($this->permissionRule()->reset);
+        if ($checkPermission !== true) {
+            $checkPermission->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($checkPermission);
+            return $this->sendResponse($checkPermission);
+        }
 
         $result = $this->repository->reset('id', $id);
+
+        if ($result->getResponseCode() !== ResponseCode::OK) {
+            $result->setRequestInfo($request, $this->repository->currentAccount()->toArray());
+            $this->saveLog($result);
+        }
         return $this->sendResponse($result);
     }
 }
